+++ library/src/main/java/com/evernote/android/job/v21/PlatformJobService.java
@@ -59,7 +59,7 @@ public class PlatformJobService extends JobService {
         final int jobId = params.getJobId();
         final JobProxy.Common common = new JobProxy.Common(this, CAT, jobId);
 
-        final JobRequest request = common.getPendingRequest(true);
+        final JobRequest request = common.getPendingRequest(true, true);
         if (request == null) {
             return false;
         }