+++ library/src/test/java/com/evernote/android/job/FailureCountTest.java
@@ -8,6 +8,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.MethodSorters;
 
+import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
@@ -53,5 +54,25 @@ public class FailureCountTest extends BaseJobManagerTest {
         jobId = job.getNewJobId();
 
         assertThat(manager().getJobRequest(jobId).getFailureCount()).isEqualTo(2);
+    }
+
+    @Test
+    public void verifyDeletedJobIsNotPersistedAgain() throws Exception {
+        int jobId = DummyJobs.createBuilder(DummyJobs.TwoSecondPauseJob.class)
+                .setPeriodic(TimeUnit.MINUTES.toMillis(15))
+                .build()
+                .schedule();
+
+        Future<Job.Result> future = executeJobAsync(jobId, Job.Result.SUCCESS);
+
+        // wait until the job is started
+        Thread.sleep(1_000);
+        assertThat(manager().getJob(jobId)).isNotNull();
+
+        // will also cancel the running job
+        manager().cancel(jobId);
+
+        assertThat(future.get(3, TimeUnit.SECONDS)).isEqualTo(Job.Result.SUCCESS);
+        assertThat(manager().getJobRequest(jobId)).isNull();
     }
 }