+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.11
+
+* Fix a race condition when canceling jobs, see #178
+
 ## 1.1.10 (2017-04-29)
 
 * Fix a race condition when canceling jobs, see #178